Add limiting scope to pom.xml as suggested in README
As the README states:

Derive4J should be declared as a compile-time only
dependency (not needed at runtime). So while derive4j is
(L)GPL-licensed, the generated code is not linked to derive4j, and thus
derive4j can be used in any project (proprietary or not).

However, the provided maven tags do not impose this requirement
potentially causing legal issues for folks who accidentally forget the
proper scope in proprietary codebases.

This PR ensures that the default pom example uses the recommended
practices.